added full logic to create suggested order from query
